stormcloud/etl/transpose_geom/requirements.txt
affine==2.4.0
aiobotocore==2.5.4
aiohttp==3.8.6
aioitertools==0.11.0
aiosignal==1.3.1
asciitree==0.3.3
async-timeout==4.0.3
attrs==23.1.0
boto3==1.22.2
botocore==1.31.17
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.0.0
cycler==0.12.1
dask==2022.4.1
dataclasses-json==0.5.7
fasteners==0.19
Fiona==1.9.3
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.5.0
GDAL==3.5.1
idna==3.4
importlib-metadata==6.8.0
jmespath==1.0.1
joblib==1.3.2
kiwisolver==1.4.5
locket==1.0.0
marshmallow==3.20.1
marshmallow-enum==1.5.1
matplotlib==3.5.1
multidict==6.0.4
munch==4.0.0
mypy-extensions==1.0.0
numcodecs==0.12.1
numpy==1.24.3
packaging==23.2
pandas==2.0.3
partd==1.4.1
Pillow==10.1.0
# pydsstools==2.2
pyparsing==3.1.1
pyproj==3.5.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2023.3.post1
PyYAML==6.0.1
rasterio==1.3.9
rioxarray==0.11.1
s3fs==2023.5.0
s3transfer==0.5.2
scikit-learn==1.2.2
scipy==1.10.1
shapely==2.0.1
six==1.16.0
snuggs==1.4.7
threadpoolctl==3.2.0
toolz==0.12.0
typing-extensions==4.8.0
typing-inspect==0.9.0
tzdata==2023.3
urllib3==1.26.18
wrapt==1.15.0
xarray==2022.3.0
yarl==1.9.2
zarr==2.14.2
zipp==3.17.0
